+++ script/tool/lib/src/build_examples_command.dart
@@ -11,6 +11,7 @@ import 'common/core.dart';
 import 'common/package_looping_command.dart';
 import 'common/plugin_utils.dart';
 import 'common/process_runner.dart';
+import 'common/repository_package.dart';
 
 /// Key for APK.
 const String _platformFlagApk = 'apk';
@@ -104,7 +105,7 @@ class BuildExamplesCommand extends PackageLoopingCommand {
   }
 
   @override
-  Future<PackageResult> runForPackage(Directory package) async {
+  Future<PackageResult> runForPackage(RepositoryPackage package) async {
     final List<String> errors = <String>[];
 
     final Iterable<_PlatformDetails> requestedPlatforms = _platforms.entries
@@ -135,9 +136,9 @@ class BuildExamplesCommand extends PackageLoopingCommand {
     }
     print('');
 
-    for (final Directory example in getExamplesForPlugin(package)) {
+    for (final RepositoryPackage example in package.getExamples()) {
       final String packageName =
-          getRelativePosixPath(example, from: packagesDir);
+          getRelativePosixPath(example.directory, from: packagesDir);
 
       for (final _PlatformDetails platform in buildPlatforms) {
         String buildPlatform = platform.label;
@@ -158,7 +159,7 @@ class BuildExamplesCommand extends PackageLoopingCommand {
   }
 
   Future<bool> _buildExample(
-    Directory example,
+    RepositoryPackage example,
     String flutterBuildType, {
     List<String> extraBuildFlags = const <String>[],
   }) async {
@@ -168,12 +169,12 @@ class BuildExamplesCommand extends PackageLoopingCommand {
     // needs to be created on the fly with 'flutter create .'
     Directory? temporaryPlatformDirectory;
     if (flutterBuildType == 'winuwp') {
-      final Directory uwpFolder = example.childDirectory('winuwp');
+      final Directory uwpFolder = example.directory.childDirectory('winuwp');
       if (!uwpFolder.existsSync()) {
         print('Creating temporary winuwp folder');
         final int exitCode = await processRunner.runAndStream(
             flutterCommand, <String>['create', '--platforms=winuwp', '.'],
-            workingDir: example);
+            workingDir: example.directory);
         if (exitCode == 0) {
           temporaryPlatformDirectory = uwpFolder;
         }
@@ -189,7 +190,7 @@ class BuildExamplesCommand extends PackageLoopingCommand {
         if (enableExperiment.isNotEmpty)
           '--enable-experiment=$enableExperiment',
       ],
-      workingDir: example,
+      workingDir: example.directory,
     );
 
     if (temporaryPlatformDirectory != null &&