+++ script/tool/test/common/plugin_utils_test.dart
@@ -6,6 +6,7 @@ import 'package:file/file.dart';
 import 'package:file/memory.dart';
 import 'package:flutter_plugin_tools/src/common/core.dart';
 import 'package:flutter_plugin_tools/src/common/plugin_utils.dart';
+import 'package:flutter_plugin_tools/src/common/repository_package.dart';
 import 'package:test/test.dart';
 
 import '../util.dart';
@@ -21,7 +22,8 @@ void main() {
 
   group('pluginSupportsPlatform', () {
     test('no platforms', () async {
-      final Directory plugin = createFakePlugin('plugin', packagesDir);
+      final RepositoryPackage plugin =
+          RepositoryPackage(createFakePlugin('plugin', packagesDir));
 
       expect(pluginSupportsPlatform(kPlatformAndroid, plugin), isFalse);
       expect(pluginSupportsPlatform(kPlatformIos, plugin), isFalse);
@@ -32,7 +34,8 @@ void main() {
     });
 
     test('all platforms', () async {
-      final Directory plugin = createFakePlugin('plugin', packagesDir,
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+          'plugin', packagesDir,
           platformSupport: <String, PlatformDetails>{
             kPlatformAndroid: const PlatformDetails(PlatformSupport.inline),
             kPlatformIos: const PlatformDetails(PlatformSupport.inline),
@@ -40,7 +43,7 @@ void main() {
             kPlatformMacos: const PlatformDetails(PlatformSupport.inline),
             kPlatformWeb: const PlatformDetails(PlatformSupport.inline),
             kPlatformWindows: const PlatformDetails(PlatformSupport.inline),
-          });
+          }));
 
       expect(pluginSupportsPlatform(kPlatformAndroid, plugin), isTrue);
       expect(pluginSupportsPlatform(kPlatformIos, plugin), isTrue);
@@ -51,12 +54,13 @@ void main() {
     });
 
     test('some platforms', () async {
-      final Directory plugin = createFakePlugin('plugin', packagesDir,
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+          'plugin', packagesDir,
           platformSupport: <String, PlatformDetails>{
             kPlatformAndroid: const PlatformDetails(PlatformSupport.inline),
             kPlatformLinux: const PlatformDetails(PlatformSupport.inline),
             kPlatformWeb: const PlatformDetails(PlatformSupport.inline),
-          });
+          }));
 
       expect(pluginSupportsPlatform(kPlatformAndroid, plugin), isTrue);
       expect(pluginSupportsPlatform(kPlatformIos, plugin), isFalse);
@@ -67,7 +71,8 @@ void main() {
     });
 
     test('inline plugins are only detected as inline', () async {
-      final Directory plugin = createFakePlugin('plugin', packagesDir,
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+          'plugin', packagesDir,
           platformSupport: <String, PlatformDetails>{
             kPlatformAndroid: const PlatformDetails(PlatformSupport.inline),
             kPlatformIos: const PlatformDetails(PlatformSupport.inline),
@@ -75,7 +80,7 @@ void main() {
             kPlatformMacos: const PlatformDetails(PlatformSupport.inline),
             kPlatformWeb: const PlatformDetails(PlatformSupport.inline),
             kPlatformWindows: const PlatformDetails(PlatformSupport.inline),
-          });
+          }));
 
       expect(
           pluginSupportsPlatform(kPlatformAndroid, plugin,
@@ -128,7 +133,8 @@ void main() {
     });
 
     test('federated plugins are only detected as federated', () async {
-      final Directory plugin = createFakePlugin('plugin', packagesDir,
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+          'plugin', packagesDir,
           platformSupport: <String, PlatformDetails>{
             kPlatformAndroid: const PlatformDetails(PlatformSupport.federated),
             kPlatformIos: const PlatformDetails(PlatformSupport.federated),
@@ -136,7 +142,7 @@ void main() {
             kPlatformMacos: const PlatformDetails(PlatformSupport.federated),
             kPlatformWeb: const PlatformDetails(PlatformSupport.federated),
             kPlatformWindows: const PlatformDetails(PlatformSupport.federated),
-          });
+          }));
 
       expect(
           pluginSupportsPlatform(kPlatformAndroid, plugin,
@@ -189,13 +195,13 @@ void main() {
     });
 
     test('windows without variants is only win32', () async {
-      final Directory plugin = createFakePlugin(
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
         'plugin',
         packagesDir,
         platformSupport: <String, PlatformDetails>{
           kPlatformWindows: const PlatformDetails(PlatformSupport.inline),
         },
-      );
+      ));
 
       expect(
           pluginSupportsPlatform(kPlatformWindows, plugin,
@@ -208,13 +214,14 @@ void main() {
     });
 
     test('windows with both variants matches win32 and winuwp', () async {
-      final Directory plugin = createFakePlugin('plugin', packagesDir,
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+          'plugin', packagesDir,
           platformSupport: <String, PlatformDetails>{
             kPlatformWindows: const PlatformDetails(
               PlatformSupport.federated,
               variants: <String>[platformVariantWin32, platformVariantWinUwp],
             ),
-          });
+          }));
 
       expect(
           pluginSupportsPlatform(kPlatformWindows, plugin,
@@ -227,13 +234,14 @@ void main() {
     });
 
     test('win32 plugin is only win32', () async {
-      final Directory plugin = createFakePlugin('plugin', packagesDir,
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
+          'plugin', packagesDir,
           platformSupport: <String, PlatformDetails>{
             kPlatformWindows: const PlatformDetails(
               PlatformSupport.federated,
               variants: <String>[platformVariantWin32],
             ),
-          });
+          }));
 
       expect(
           pluginSupportsPlatform(kPlatformWindows, plugin,
@@ -246,14 +254,14 @@ void main() {
     });
 
     test('winup plugin is only winuwp', () async {
-      final Directory plugin = createFakePlugin(
+      final RepositoryPackage plugin = RepositoryPackage(createFakePlugin(
         'plugin',
         packagesDir,
         platformSupport: <String, PlatformDetails>{
           kPlatformWindows: const PlatformDetails(PlatformSupport.federated,
               variants: <String>[platformVariantWinUwp]),
         },
-      );
+      ));
 
       expect(
           pluginSupportsPlatform(kPlatformWindows, plugin,