+++ script/tool/test/lint_android_command_test.dart
@@ -101,7 +101,7 @@ void main() {
           });
 
       processRunner.mockProcessesForExecutable['gradlew'] = <io.Process>[
-        MockProcess.failing(),
+        MockProcess(exitCode: 1),
       ];
 
       Error? commandError;