+++ packages/in_app_purchase/in_app_purchase_android/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 0.1.4+5
 
+* Add `implements` to pubspec.
 * Updated Android lint settings.
 
 ## 0.1.4+4